+++ src/sdl3/render.c
@@ -7,7 +7,7 @@ i8 _wh_window_create_sdl3(_wh_window_create_params params) {
 	wh_graphics_s* grap = &params.ins->graphics;
 	wh_log_debug(("Starting SDL3!"));
 
-	grap->window.sdl = SDL_CreateWindow(params.title.str, params.size_x, params.size_y, 0);
+	grap->window.sdl = SDL_CreateWindow(params.title.str, (int)params.size_x, (int)params.size_y, 0);
 	grap->sdl3.renderer = SDL_CreateRenderer(grap->window.sdl, NULL);
 
 	return 0;
@@ -41,7 +41,7 @@ void _wh_render_line_sdl3(_wh_render_line_params params) {
 	SDL_FPoint* points = (void*)params.data;
 	params.count /= 2;
 
-	if (false == SDL_RenderLines(params.ins->graphics.sdl3.renderer, points, params.count)) {
+	if (false == SDL_RenderLines(params.ins->graphics.sdl3.renderer, points, (int)params.count)) {
 		wh_log_error(("Failed to render line! [ %s ]"), SDL_GetError());
 	}
 }
